Added support for ExportPermissionsFile
The new parameter ExportPermissionsFile, requires some minor changes in way how this scripts was yours works.
In the RESTORE_PERMISSIONS, the way of T-SQL commands as generated has changed. Now, it get the T-SQL DCL commands from same place of that code that will export. This avoids different commands in export and the executed after restore.

In the DESTINATION_PERMISSION_BACKUP, the implementation of the parameter was added.